-Renamed the package name -Added exclusions to gitignore
