+++ src/main/java/com/github/jenkinsci/zaproxyplugin/report/ZAPreportXML.java
@@ -22,26 +22,25 @@
  * SOFTWARE.
  */
 
-package fr.novia.zaproxyplugin.report;
+package com.github.jenkinsci.zaproxyplugin.report;
 
-import org.zaproxy.clientapi.core.ClientApi;
 import org.zaproxy.clientapi.core.ClientApiException;
+import org.zaproxy.clientapi.core.ClientApi;
 
 /**
- * Used to generate ZAP report in html. 
+ * Used to generate ZAP report in xml. 
  * 
  * @author ludovic.roucoux
  *
  */
-public class ZAPreportHTML extends ZAPreport {
+public class ZAPreportXML extends ZAPreport {
 
-	public ZAPreportHTML() {
-		this.format = ZAPreport.REPORT_FORMAT_HTML;
+	public ZAPreportXML() {
+		this.format = ZAPreport.REPORT_FORMAT_XML;
 	}
 
 	@Override
 	public byte[] generateReport(ClientApi clientApi, String apikey) throws ClientApiException {
-		return clientApi.core.htmlreport(apikey);
+		return clientApi.core.xmlreport(apikey);
 	}
-
 }
